restructured placement of hooks
