+++ packages/protocolbeat/src/apps/discovery/defidisco/DeFiScanPanel.tsx
@@ -5,7 +5,7 @@ import { getFunctions, getFundsData, getProject } from '../../../api/api'
 import type { ApiFundsDataResponse } from '../../../api/types'
 import { Checkbox } from '../../../components/Checkbox'
 import { V2ScoringSection } from '../../../defidisco/V2ScoringSection'
-import { useContractTags } from '../../../hooks/useContractTags'
+import { useContractTags } from './hooks/useContractTags'
 import { usePanelStore } from '../store/panel-store'
 import { FundsSection } from './FundsSection'
 import { formatUsdValue } from './formatUtils'
