+++ packages/protocolbeat/src/apps/discovery/defidisco/ExternalButton.tsx
@@ -3,8 +3,8 @@ import { useParams } from 'react-router-dom'
 import {
   useContractTags,
   useUpdateContractTag,
-} from '../../../hooks/useContractTags'
-import { useExternalToggle } from '../../../hooks/useExternalToggle'
+} from './hooks/useContractTags'
+import { useExternalToggle } from './hooks/useExternalToggle'
 import { ControlButton } from '../panel-nodes/controls/ControlButton'
 import { useStore } from '../panel-nodes/store/store'
 import { DependencyPropagationDialog } from './DependencyPropagationDialog'
