+++ packages/protocolbeat/src/apps/discovery/panel-list/ListPanel.tsx
@@ -7,7 +7,7 @@ import type { ApiAddressEntry, ApiProjectChain } from '../../../api/types'
 import { AddressIcon } from '../../../components/AddressIcon'
 import { ErrorState } from '../../../components/ErrorState'
 import { LoadingState } from '../../../components/LoadingState'
-import { useContractTags } from '../../../hooks/useContractTags'
+import { useContractTags } from '../defidisco/hooks/useContractTags'
 import { IconChevronDown } from '../../../icons/IconChevronDown'
 import { IconChevronRight } from '../../../icons/IconChevronRight'
 import { IconFolder } from '../../../icons/IconFolder'
