+++ packages/protocolbeat/src/apps/discovery/panel-nodes/view/NodeView.tsx
@@ -2,7 +2,7 @@ import clsx from 'clsx'
 
 import { AddressIcon } from '../../../../components/AddressIcon'
 import { IconInitial } from '../../../../icons/IconInitial'
-import { useIsContractExternal } from '../../../../hooks/useContractTags'
+import { useIsContractExternal } from '../../defidisco/hooks/useContractTags'
 import type { Field, Node } from '../store/State'
 import { useStore } from '../store/store'
 import {
